Make the parameter for addcommas a long
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1150 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/centipede.d/stuff.c
@@ -68,11 +68,11 @@ displaymen()
 }
 
 char *addcommas(num)
-int num;
+long num;
 {
     static char str[14];
     register char *spot = str;
-    register int p10;
+    register long p10;
 
     for (p10 = 1000; num / p10 > 0; p10 *= 1000);
     p10 /= 1000;
